requirements.txt
flask==3.0.0
moviepy==1.0.3
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
torch==2.1.1
tqdm==4.66.1
werkzeug==3.0.1
yt-dlp==2023.12.30
openai==0.28
ffmpeg-python==0.2.0
opencv-python==4.9.0.80
tiktoken==0.5.2
huggingface-hub
tokenizers
transformers
faster-whisper
youtube-transcript-api
gunicorn==21.2.0
